Remove MacOS from GitHub Actions
We only really care that this works on different Perl versions.

We have CPAN Testers for different platforms. There is no platform-
specific code here, and if there is a problem on MacOS or Windows
etc then it is likely due to issues with prerequisites than this
module.

There's no reason to waste CPU cycles and fuel running tests on
other platforms that we don't expect there to be a problem with.